Pull illumos bits into a module to make clippy happy
